Some systems can not tell the diffrence from String and string rename String to JString
