+++ JString.cpp
@@ -24,7 +24,7 @@
 */
 #include <assert.h>
 
-#include "String.h"
+#include "JString.h"
 
 namespace json
 {
