+++ JString.h
